combine pipeline/batch files and put into subdirectories
The test list is always stored in "testlist", then when vvtest runs,
"launched" and "finished" files are produced.  This is fine, but in
pipeline mode there will be a testlist.# and launched.# and finished.#
for each pipe.  This can add up to a lot of files.  This commit
combines the testlist, launched, and finished into just "testlist".

In addition, all the pipeline/batch files are now placed into specially
named subdirectories of TestResults.  Each subdirectory will contain
at most 50 batches.  They are named "batchset0", "batchset1", etc.

Also, add a separate batch handling class called ProcessorBatch for
running in batch mode but on a regular UNIX machine.  This helps test
the pipeline mode without running on a batch system.

Also, move the PermissionSetter class into its own file in libvvtest.