add -G option to results.py report
The -G <file glob pattern> will do the same thing as -g except that
the tests that diff/fail/timeout will not be detailed in the test
detail section.  This option will be used to report the results of a
new platform/compiler without flooding the test detail section.

Tests: unit tests PASS