move certain logic out of the test script
..and into the python coding.

Right before a test is executed, the files in the execute directory are
removed (except for a few vvtest files).  This is to avoid possible issues
with left over files if the test is run more than once.  This removal was
taking place in the test run script, now it happens right before the test
script is executed.

When a test completes execution and the -C option is given to vvtest, most
files in the execute directory are removed.  This was being done in the
test script, now it is done in the python code just after the test script
returns.

Change the mechanism for glob'ing files.  Instead of doing it at test parse
time, it is now done right before the test is executed.  Or in the case of
--extract, it is done during the file copying.

Deprecate the <glob_link> and <glob_copy> XML elements, which were probably
never used.

Remove unused PRERUN and POSTRUN mechanisms.

Tests: unit tests PASS