Add hook for parsing "*.vvt" test files
If a file with the vvt extension is found in the scan, the code for
parsing
and incorporating test instances will be called.  Nothing is done in
there
for now.  Baby steps!

Move XML document parser into TestSpecCreator.py to encapsulate all the
XML
parsing into that file.

Tests: unit tests PASS