first working general script as a test
This commit modifies a few things and prepare for specifying tests
by using a file that is just a script.  There are a number of places
that still have to be fleshed out, but a simple executable script
can now be run by vvtest.  The file name has to have the extension
".vvt".

Tests: unit tests PASS