fix parameterize bugs and preclean function bug
With execute/analyze tests, the analyze test receives the list
of parameters that was just executed in the child tests.  But
when using the -p option to restrict the tests that get run
to certain parameter values, the analyze test was not receiving
the correct list of parameter values.

Also, in the preclean() function, soft linked directories were
being sent into shutil.rmtree() when they should be just os.remove().

Tests: unit tests PASS