adjust --save-results file date stamp
When --save-results is on the command line of a vvtest sequence,
a test results file is written right before the tests start to
run and overwritten after the test sequence finishes.  Sometimes in
the long tests in the nightlies, the date stamp was rolling over to
the next morning and so the first file would not be overwritten with
the final results.  This would cause the results reporting to show
that a test sequence was started but never finished.

This issue should now be fixed.  The date stamp when the test list file
is first written is saved and when the final results are written, that
same date stamp is used.

Tests: all unit tests PASS