add TestResults "locking"
Multiple instances of vvtest cannot be running the tests in the same
TestResults directory at the same time.  This commit adds a sort of
"lock" marker in the testlist file in order to indicate that a
process is currently running in that TestResults directory.  The
main reason is to avoid accidental running of another vvtest which
can corrupt the working files in there.  If a second vvtest is launced
an error message is printed and it exits.  If a vvtest dies badly and
there appears to be an existing vvtest running, a second vvtest can be
forced to run by adding the --force option.

This also changes the -F to -R instead in order to avoid confusion
with the "force" option.  Now -R means "rerun".  The -F option is
deprecated but still works as before for now.

Tests: unit tests PASS