add initial test util script fragment capability
For tests using a general script file, a vvtest utility script fragment
adds a lot of convenience for test writers and maintainers.  This commit
adds the logic hooks for writing script fragments into the test
execution directories and fills out a fragment when the script language
is python.

Tests: unit tests PASS, added unit test for the new capability