fix issues with --pipeline and the finished mark
There was a subtle behavior issue with the "finished" date stamp
that is written to the test list file.  In --pipeline mode, it was
not getting written, but a second oversight prevented a failure
to occur when the stamp was not written.  Both of these issues are
now fixed.

Tests: unit tests PASS, another pipeline test added