adjust test script util
The vvtest_util.* is written for each test.  This script was always
including the libvvtest/script_util.*.  Now it looks for a file
called script_util_plugin.* from the config directory, and if that
exists, it only includes that.  If it does not, then the libvvtest
one is included instead.

Another change is that the variables defined in the platform
plugin are made available to the plugin util script, and only
applied to the environment if there is not a plugin util script.

Tests: unit tests PASS, added another config one