for script tests, add handling of -a, --analyze
For sh/bash test scripts, save off the test script arguments into
variables
so functions can access them in the function utilities if they want.

In the test scripts utilities, added parsing of
--execute_analysis_sections
and set a flag.  This option gets sent into tests if the -a or --analyze
option is given on the vvtest command line.

Add two new functions: analyze_only() for both python and bash, and
cmdline_option() for bash.

Tests: PASS tests and add another