no longer cache the command line keywords
The keywords given on the command line with -k and -K were stored in
the TestResults.*/test.cache file, then merged in with any new command
line keyword expressions when vvtest is run again.  This no longer
happens.  The keywords are not stored, so each time vvtest is run, the
keyword expression is exactly what was specified on the command line.

Hopefully this solves much of the dreaded "no tests to run" issue.

Tests: unit tests PASS