fix incorrect logic when a batch system does not specify maxprocs
On batch systems that do not specify a max number of processors in their
config, the number of processors on the login node was being used.  The
workaround is to use -N 512 or some big number, but this commit makes
the max in this case be 2**31-1 (that is, no maximum).