add html test results rollup reporting capability
Test results reporting has had a text form of a report.  Now it has an
html report form.  This is in the results.py file.

Also in this commit is a small change to allow options to be passed
from a platform plugin to the batch system implementation.

Tests: unit tests PASS