add more robustness & flexibility to idplatform.py plugin
The file idplatform.py a config directory is used to identify the name
of the platform and of the compiler.  With this commit, either function
(platform and compiler) can be missing and vvtest will use a default.
Also, if either function returns None or an empty string, the default
is used for the value.

Tests: unit tests PASS; add coverage for this logic