add TDD test marking
If the keyword "TDD" is added to a test, then that test will not be run
by default.  To include tests that are marked TDD, add the --include-tdd
option to the vvtest command line.

This feature is intended to allow for tests that are not yet working but
can be added to the test suite without causing diffs/fails to show up
in the production test reporting.