make runcmd.py and runjob.py complimentary capability
Processes that run commands need to run some in the foreground, some in
the background, and some on remote machines.  The runcmd.py and runjob.py
utilties are now compliments, where the runcmd.py functions run in the
foreground, and runjob.py runs jobs in the background on the same machine
or on a remote machine.