trigger: add 'name' attribute to run_job()
The 'name' attribute is cached in the Job class.  It can be used to help
manage your jobs, but the main reason for the addition is that the
'name' is used as the prefix to the log file name (the default is to
use the basename of the program being run.