add remote object persistence utility
For the remote python execution capability, this commit adds three functions
that provide a way for objects created on the remote side to persist and be
passed between the remote and local sides.  Basically, the id() of remote
objects are stored in a global map object (on the remote side), and that id
is passed back and forth.

This method is now used in the runjob.py coding to open a file on the remote
side, read from that file incrementally, and close the file.