Corrected NP filtering
When vvtest -N is used on an execute/analyze test, the whole test is skipped
if even one of the child tests ask for too many processors.  This commit fixes
the case where there is NOT an analyze.  Without an analyze, only the tests
that exceed the -N value should be skipped.  There was a bug that caused all
tests to be skipped anyway.

Changed function name loadExecutionTimes to loadRunTimes, and move the call to
it up before loadAndFilter().  This is in preparation to filtering tests based
on the previous runtime.  Also added loadRunTimes to other functions that
construct the test list and do filtering.

Copy the runtests script from trig for the added features: running multiple
tests at one time, and iterating over python versions.