refactor TestList's storage of TestSpec objects
Instead of two lists, scantests and filetests, now there is just
one dictionary called tspecs.  The two lists were a way of keeping
track of whether a test object came from a string (a test list file)
or from the source test file (a .xml or .vvt).  To accomplish the
same thing in a different way, the TestSpec class now has an "origin"
data member.  It is a list with values like "file", "string", or "copy".
Because it is a list, a sort of history of the test modifications are
kept.