trigger: add filecopy.py utility
The filecopy.py utility copies files and directories between two machines.
Like filesync.py, it has a command line entry point and a function entry
point that can be used easily when importing as a python module.

This also changes an option for the filesync.py command line.  The -p option
has been replaced with the -g option (for glob).