add waitforjobid attribute to runjob.run_job()
Background jobs can now be chained, in that a job can be made to only start
after some other job completes.  A limitation is that a job can only depend
on one other job.  Add waitforjobid=<jobid> as a keyword argument in the
runjob.run_job() function.

Some refactoring of runjob.py was done to improve the code and to accommodate
the waitforjobid capability.