trig: add timeout_date attribute to runcmd and runjob
There is an existing method to timeout a command, timeout=<num seconds>.  This
commit adds the timeout_date=<epoch seconds> attribute to the
runcmd.run_timeout() and runjob.run_job() commands.  The value <epoch seconds>
is expected to be a date, such as that given by time.time(), but would be a
date into the future, such as tommorow at 5pm.