trig: add 'sharedlog' attribute to runjob
When running remote jobs that share an NFS file system, it is sort of a
waste to have the remote log be produced and copied back to the local
log directory.  In this case, the sharedlog=True option can be passed to
a runjob.run_job() function, which causes the remote process to redirect
its output directly to the local log file name.