trig: allow waitforjobid=None in runjob.run_job()
Consider a use case like this for running conditional jobs in sequence:

  if day_of_week == 'Mon':
    jid = run_job( cmd1 )
    wait_job( jid )
  if day_of_week == 'Tue':
    jid = run_job( cmd2 )
    wait_job( jid )
  if day_of_week in ['Tue','Thu']:
    jid = run_job( cmd3 )
    wait_job( jid )

While this works, there are some (minor) drawbacks.  One is the logging; the
log output does not show all jobs that will run until they actually run.

An easy change to the 'waitforjobid' attribute can help.  It can now take on
the value of None.  This allows the use case to be written:

  jid = None
  if day_of_week == 'Mon':
    jid = run_job( cmd1, waitforjobid=jid )
  if day_of_week == 'Tue':
    jid = run_job( cmd2, waitforjobid=jid )
  if day_of_week in ['Tue','Thu']:
    jid = run_job( cmd3, waitforjobid=jid )
  wait_all()

To be clear, a value of None is ignored - it is treated like waitforjobid was
never specified at all.