more complete filtering of baseline tests
In some cases when filters were applied to rebaselining specifications,
the test would still show as executing the rebaseline operation even
though nothing was being done for the test.  This commit causes those
tests for which no operation would be done to be filtered out so they
do not show up as being rebaselined.