modernize the xmlwrapper implementation
The xmlwrapper module has been very effective over more than a decade in
providing XML parsing for the (simple) test description files.  The module
defines a small class as an interface to the XML document, so as to hide
the implementation.

More recently, the xml.etree.ElementTree module is available for Python 2.5
and greater.  Using it reduces the amount of code to implement the wrapper
interface.

Also got rid of raw data member accesses into the wrapper class.

Also added unit testing for the module.  It doesn't cover all features
but most of them.

Also ported the module and testing to Python 3.