remove script_util_plugin from default
A project can import helper python or shell scripts from the config
directory.  A file called script_util_plugin was placed in the vvtest
area so that tests could import that whether or not a project had
one of that name.  However, this is not really necessary and only
adds confusion when trying to understand vvtest operations.