add environment bootstrap mechanism for vvtest
Vvtest can use a project config directory that contains script
utility helpers.  However, doing this requires an environment
variable or a command line option to be used.  This commit adds
a bootstrap mechanism that uses the location of vvtest (or a
soft link to vvtest) to execute a vvtest_bootstrap.py file.  This
file can then set environment variables and/or add to the sys.path
list.  A test can then count on the variables existing or import
pathing to be set for the platform.