modify bootstrap mechanism and move script utils
This moves the script_util files into the config directory, to make
it easier to find them as test writers.

The bootstrap file is now imported before the config directory is
determined.  This allows the bootstrap file to specify VVTEST_CONFIGDIR
environment variable, which then is used to set the config directory.