vvtest: remove some uses of TestSpec.getParent()
This is a refactor centered around removing the use of the "parent" and
"child" concepts.  The TestSpec.getParent() function was being used to check
if a test is an analyze test in a parameterize/analyze group.  This commit
removes many of the uses of getParent() used in this way.  This required
changes to how locations in the code ask if a test is an analyze test, and
therefore how analyze specifications are handled.

There are a few more uses of getParent() that need to be removed.