vvtest: remove set/getParent() from TestSpec
Refactoring things so the use of getParent() is no longer needed.  This
required an augmentation to the test list files used in batch mode.  They
now contain some dependency information.

The main need to know dependencies (currently) is that the parameterized
tests cannot clean themselves out when the -C option is given (which is
the post clean option).  Their files may be needed by the analyze test.