All vvtest-related imports are now relative.
This change was motivated by me trying to use h5py in a .vvt file
but it would fail. After some digging I realized that it was because
the libvvtest directory was prepended to the PATH, which made the
libvvtest/platform.py take precendence over the system module by
the same name.

Beyond this one issue, it's generally safer to do relative imports
so that things like this don't keep popping up.

All expected tests pass on ceesrv02.