remove sys.path modification in python test utils
For python script tests, the sys.path was being modified in the
script util file.  It was there to support the import of a filtering
utility within the main libvvtest directory.

This commit does remove the sys.path modification.  However, the
PYTHONPATH was added to just prior to invoking python script tests.
So really the sys.path is still modified, just in a different place.
But the directory added to PYTHONPATH is different - it is the path
containing the vvtest executable, which is consistent with using
python package imports rather than modifying sys.path then doing
a plain import.

Another minor modification in this commit: a variable called CONFIGDIR
is now always added to script util files.  This may be needed by
advanced script usage, and is currently used in the bash/sh util file.