make TestExec.getBlockingTest always correct
There was a function in TestList to get blocking dependencies that
used information in the TestList in order to be correct in all cases.
The getBlockingTest function of TestExec class could be made to
be correct in all cases too by adding one line to TestExec ctor.
When constructing a TestExec, the test "state" had to be marked
as not run.

After this change, the TestList get blocking dependencies function
could be removed.