add dependency result handling
A "depends on" directive can now give an expression for whether or not
the dependency result should allow the test to run.  For example, this

  VVT: depends on (result="*") : testY

would allow the test to run no matter what the result value of testY
was.  A general expression is allowed, such as "not fail" or "pass or
diff" or "pass".