change dependency DEPDIRMAP to multi-valued
A test that "depends on" another test specified as a pattern (like
glob-ing) can result in more than one value (more than one test).
The vvtest_util.py file in the test directory contains the DEPDIRMAP
which now maps the pattern given in the "depends on" specification
to a list of test execution directories.