remove use of "non results" keywords
Keywords are general strings but also include the results of a test,
such as notrun, notdone, pass, fail, diff.  The -k and -K options
are used to create expressions that are evaluated against the keywords
in a test.

At one point, there was a need to evaluate expressions with the results
keywords removed from the expression.  This meant more coding and was a
little tricky.  I removed the special non-results keyword expression
evaluation and all tests passed.  I also beefed up the main test of this
use case.

Since all tests passed, I'm going with it.