fix obscure ordering bug in XML test syntax
When "zipping" parameters together in the XML test specification syntax,
such as

  <parameterize np="1 4" dt="0.1 0.01"/>

sometimes the order of the combined variable, such as PARAM_np_dt,
in the helper scripts was not consistent (the same order as in the
specification).  This commit sorts the zipped parameter names so
that the order is always consistent (it is alphabetic).