change BatchJob processor/node set and get methods
The processor allocation is different for different queuing systems - some
want num nodes (eg SLURM) and some want num CPUs (eg LSF).  The tie between
them is the number of processors per node, but that is a platform configuration
thing.  This commit adds a platform configuration class that knows the machine
and queue specifications (such as ppn).  The processor specification to a
batch job is also changed to include both num cores and num nodes, which
allows the batch implementation to pick the one it needs.