fix DEPDIRS bug with analyze reruns
If an analyze test fails (but the children pass), then the analyze test
is rerun using vvtest -k fail, the DEPDIRS in the test should still
contain the children paths.  It was not; now it is.  Added test to
suite that failed before and passes now.

Issue: https://gitlab.sandia.gov/rrdrake/scidev/issues/9