add ram disk subversion password caching script
The svnpasscache script writes a subversion config directory into
a RAM disk protected by unix permissions.  The svnwrap script can
then be used to perform subversion operations without entering a
password.

The gnome-keyring method turns out to be too problematic to be
viable (that is, I can't get it to work for an entity account).

This does store the password in plain text in memory, but at least
it does not write it to disk.  While more obfuscated, gnome-keyring
essentially does the same thing.