refactor test launch command construction
The coding to generate the actual subprocess command used to execute
each test was complicated and ugly.  This commit refactors it into a
much cleaner mechanism.

A couple of features have been sacrificed to achieve this, which are
explained in the RELEASE_NOTES.  I very much doubt anyone really used
or needed them, however.

The goal here is to prepare for added features:

  1. Put the tests in their own process group so that kill signals will
     be sent to all children processes of the test
  2. Provide a per-test environment setup mechanism