more reliable test process killing
This commit changes the way the test subprocesses are killed (sent signals).
It uses a UNIX process group in order to send a signal to all child processes
at one time.  However, it was necessary to introduce an intermediary process
that catches Ctrl-C in the same group as vvtest is running in.  This process
just waits on the actual test, and dispatches signals to the test's process
group.