remove use of background processes in kill tests
The tests that kill vvtest then run it again and check behavior
was using a background process.  This, however, does not work well
on NFS because files can't be removed sometimes if a process still
has it open.

This commit changes those tests to run vvtest for a few seconds
then signal it to force it to abort.