add some robustness for testing on NFS file systems
When processes have open files on NFS mounts, you can have problems
deleting directories due to ".nfs*** device busy" type errors.  This
commit adds a utility function to try hard to remove paths and if it
is not successful, then rename it to a different filename.

This allows all vvtest unit tests to run on an NFS drive.