add use case test then fix issue
If a batch run is killed before the result file includes are inlined,
the only record of the results exist in the batch directories.  But
the batch directories are cleaned out before running the next time.
This means that previous results will be lost if they are not inlined
during vvtest startup.  A test was written to expose this case, then
the code was enhanced to fix it.