refactor the unit test hook method
There are hooks in the main execution code path that will look at
environ variables and signal the running process at the right time.
This is used for testing fault conditions in the integrated tests.

This commit refactors the way that is done, including the way the
testing is done to reduce duplication.