change default of command.py raise_on_error
The default for 'raise_on_error' in the command.py run methods is now
True.  While this is not the default for shell commands, most of the
time it is what you want.

Also, in trying to streamline the trigger.py capability, some
unnecessary options were removed from the trigger.py script.

Lastly, some test refactoring.