move vvtest specific unittest.setUp to vvtestutils.py
The vvtestutils.py module now has a small class derived from
unittest.TestCase that defines setUp() and tearDown().  The
implementations are specific to vvtest testing.  The effect
to individual tests is that the setUp() method does not need
to be defined, unless it needs to do something different than
the default.  And in that case, it just defines setUp() and
does the special stuff.