add arg parse util and populate for conversion
An argutil module was added that just imports argparse but also
provides a transparent adaptor for optparse, which is used with
python 2.6, 3.0, and 3.1.  A better help formatter was written
that allows the help description to contain paragraphs and verbatim
sections.

Add cmdline module that contains all the options and help page
documentation for vvtest command line.  Still need to convert the
rest of vvtest to use this module.