fix junit output when no tests run
This fixes a behavior with junit output when there are no tests to
run.  Previously, no junit file would be written in this case even
when --junit option was given.

Also in this commit is a refactor of how the logic for html, console,
and junit output is handled.  It has been encapuslated and put into
the libvvtest directory.