for --gitlab option, add handling of large tables
Large tables in the summary file makes it harder to scroll around to
see all the information.  This commit puts large tables in their own
file and links to that file in the summary.

If a lot of tests suddenly fail or diff, then storing all of the test
execution files can eat up a huge amount of space.  This commit limits
the number of tests that have their files stored in any given table.