workaround issue PYTHONPATH and colons
The path separator for PYTHONPATH is a colon, which creates a problem
if you want to add paths that contain colons.  For vvtest, if the
directory you are running the tests contains a colon, the tests will
not be able to import vvtest_util.py.

This commit adds a workaround for this case, by

 1. make sure your test imports script_util before vvtest_util
 2. script_util/__init__.py adjusts the sys.path

The adjustment uses the original PYTHONPATH set by vvtest and the fact
that vvtest only adds absolute paths to PYTHONPATH.