add test_preload() user plugin function
if the file vvtest_user_plugin.py file is in the config
directory and contains the function definition

def test_preload( specs ):
    pass

then it will be called just prior to each test execution.
Modifications it makes to the environment will be felt by
the test.  Also, if the test script is NOT executable and
this function returns a string, the test will be executed
using that string as the program (that is, the return value
is assumed to be a python executable to run).  This allows
the test to be run with a different version that what vvtest
is currently running as.

Lastly, each test can define a preload label, for example

  #VVT: preload : somelabel

In this case, the 'specs' dictionary argument to test_preload()
will contain the entry

  'preload': 'somelabel'

This allows each test to influence what the preload function
does.