avoid implicit keywords in user plugin functions
In the 'specs' dict passed into user plugin functions, the keywords
used to contain implicit values, such as the test name and the
parameter names.  After this change, only the keywords listed in
the test are set in the keywords list.

Also, add the test name to the 'specs' dict, as 'name'.