better fix for colon in path issue
if your tests live in a directory that contains a colon, then the
PYTHONPATH environment variable doesn't work work right.  this issue
was "fixed" previously by massaging the sys.path in script_util. the
current commit fixes it in a better way.  this fix will also allow
vvtest to be installed and run out of a directory containing a colon
(not that you would ever do that :)