add scidev/trig to vvtest dependencies
There is scidev/vvt and scidev/trig in the scidev repository.  I
want to start sharing python files/modules to avoid duplication.
This change adds the trig directory to the sys.path in a way that
supports all use cases (such as vvtest installs versus running it
out of a source code checkout).

The perms.py file from trig is now being imported by the permsetter.py
file in libvvtest.