add resilience to results git push
In the process of checking out and modifying a results branch,
another commit could be snuck into the repository.  This means
a pull has to be done before a push.  This commit adds a retry
loop which will pull then try to push again until it succeeds
(but up to 3 tries).