improvement for staged test dependencies
Adding a dependency on a staged test is now a little more intuitive.
You can add "mytest.np", for example as a "depends on" pattern.  It
will find and depend on the last stage.

If you really do want to depend on an intermediate stage, you can
do so by using something like "mytest.np*stage=1*".